Move SearchAssistVisibility to settings-api
